Frontend/requirements.txt
opencv-python-headless
python-socketio[client]
websocket-client
numpy
